debt - Avoid Number.MIN_VALUE and Number.MAX_VALUE
