+++ src/vs/editor/common/modes/linkComputer.ts
@@ -5,7 +5,6 @@
 
 import { CharCode } from 'vs/base/common/charCode';
 import { CharacterClassifier } from 'vs/editor/common/core/characterClassifier';
-import { Uint8Matrix } from 'vs/editor/common/core/uint';
 import { ILink } from 'vs/editor/common/modes';
 
 export interface ILinkComputerTarget {
@@ -32,6 +31,32 @@ export const enum State {
 }
 
 export type Edge = [State, number, State];
+
+export class Uint8Matrix {
+
+	private readonly _data: Uint8Array;
+	public readonly rows: number;
+	public readonly cols: number;
+
+	constructor(rows: number, cols: number, defaultValue: number) {
+		const data = new Uint8Array(rows * cols);
+		for (let i = 0, len = rows * cols; i < len; i++) {
+			data[i] = defaultValue;
+		}
+
+		this._data = data;
+		this.rows = rows;
+		this.cols = cols;
+	}
+
+	public get(row: number, col: number): number {
+		return this._data[row * this.cols + col];
+	}
+
+	public set(row: number, col: number, value: number): void {
+		this._data[row * this.cols + col] = value;
+	}
+}
 
 export class StateMachine {
