+++ src/vs/editor/common/viewModel/prefixSumComputer.ts
@@ -3,7 +3,7 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { toUint32 } from 'vs/editor/common/core/uint';
+import { toUint32 } from 'vs/base/common/uint';
 
 export class PrefixSumIndexOfResult {
 	_prefixSumIndexOfResultBrand: void;
