+++ src/vs/editor/contrib/referenceSearch/referencesModel.ts
@@ -15,6 +15,7 @@ import { Location, LocationLink } from 'vs/editor/common/modes';
 import { ITextModelService, ITextEditorModel } from 'vs/editor/common/services/resolverService';
 import { Position } from 'vs/editor/common/core/position';
 import { IMatch } from 'vs/base/common/filters';
+import { Constants } from 'vs/base/common/uint';
 
 export class OneReference {
 	readonly id: string;
@@ -72,7 +73,7 @@ export class FilePreview implements IDisposable {
 		const { startLineNumber, startColumn, endLineNumber, endColumn } = range;
 		const word = model.getWordUntilPosition({ lineNumber: startLineNumber, column: startColumn - n });
 		const beforeRange = new Range(startLineNumber, word.startColumn, startLineNumber, startColumn);
-		const afterRange = new Range(endLineNumber, endColumn, endLineNumber, Number.MAX_VALUE);
+		const afterRange = new Range(endLineNumber, endColumn, endLineNumber, Constants.MAX_SAFE_SMALL_INTEGER);
 
 		const before = model.getValueInRange(beforeRange).replace(/^\s+/, '');
 		const inside = model.getValueInRange(range);
